requirements.txt
anthropic>=0.7.0
python-dotenv>=0.19.0
pillow>=9.0.0
pandas>=1.3.0
openpyxl>=3.0.0
PyQt6>=6.4.0